requirements.txt
anthropic==0.45.0
matplotlib==3.10.0
numpy==2.2.2
openai==1.60.0
opencv_python==4.10.0.84
Pillow==11.1.0
python-dotenv==1.0.1
scipy==1.15.1
